+++ src/cerberus/models/__init__.py
@@ -1,4 +1,5 @@
 from .asap import ConvNeXtDCNN
+from .biasnet import BiasNet
 from .bpnet import BPNet, BPNet1024, BPNetLoss, BPNetMetricCollection
 from .dalmatian import Dalmatian
 from .gopher import GlobalProfileCNN